# docs/openapi.yaml
openapi: 3.0.3
info:
  title: be_saham_chi API
  version: 0.1.0
  description: API for stock portfolio tracking, commodity market quotes, and portfolio cash management.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/public/auth/login:
    post:
      tags:
        - Auth
      summary: Login and get a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/private/portfolios:
    get:
      tags:
        - Portfolios
      summary: List portfolios
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User portfolios.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Portfolios
      summary: Create portfolio
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioRequest"
      responses:
        "201":
          description: Portfolio created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or bearer token is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated user is not allowed to access this resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    StatusOK:
      type: string
      enum:
        - ok
    StatusNOK:
      type: string
      enum:
        - nok
...
    ErrorResponse:
      type: object
      required:
        - status
        - error
      properties:
        status:
          $ref: "#/components/schemas/StatusNOK"
        error:
          oneOf:
            - type: string
            - type: object
          example: invalid request body
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: secret123
...
    LoginResponse:
      type: object
      required:
        - status
        - token
        - token_type
        - expires_at
        - user
      properties:
        status:
          $ref: "#/components/schemas/StatusOK"
        token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserBrief"
    UserBrief:
      type: object
      required:
        - user_id
        - email
        - role_id
        - status
        - rules
        - created_at
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        role_id:
          type: integer
        status:
          type: boolean
        rules:
          type: array
          items:
            type: string
          example:
            - market.gold.read
        created_at:
          type: string
          format: date-time
...
    PortfolioRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Growth
...
    PortfolioResponse:
      type: object
      required:
        - status
        - portfolio_id
        - user_id
        - name
        - base_currency_code
        - is_main
        - created_at
        - updated_at
      properties:
        status:
          $ref: "#/components/schemas/StatusOK"
        portfolio_id:
          type: string
        user_id:
          type: string
        name:
          type: string
          example: Growth
        base_currency_code:
          type: string
          example: IDR
        is_main:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PortfolioItem:
      type: object
      required:
        - portfolio_id
        - user_id
        - name
        - base_currency_code
        - is_main
        - created_at
        - updated_at
      properties:
        portfolio_id:
          type: string
        user_id:
          type: string
        name:
          type: string
          example: Growth
        base_currency_code:
          type: string
          example: IDR
        is_main:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PortfolioListResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          $ref: "#/components/schemas/StatusOK"
        data:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
...